+++ CHANGELOG.md
@@ -4,6 +4,9 @@
   static analysis and prevent bugs where operation value is always assumed to be a String object. 
   If you get analyzer errors after this change you must add an explicit type cast on the
   operation value to the desired type (String or the type of an embed).
+* Added `dataDecoder` parameter to `Delta.fromJson` and `Operation.fromJson`. Can be used to convert
+  raw JSON values to concrete data types. Mainly useful with embedded data. Text data should always
+  be `String`. 
 
 ## 1.1.3